Add CountDocumentsAsync to IRepository
Add unit tests for MongoRepository
Fix return value of DeleteAllAsync
Change MongoRepository to use primary constructor

+++ test/SIL.XForge.Scripture.Tests/Services/SFProjectServiceTests.cs
@@ -1600,7 +1600,7 @@ public class SFProjectServiceTests
         Assert.IsTrue(env.GetProject(projectId).UserRoles.ContainsKey(userToRemove));
 
         // Delete the project user config
-        int deleted = await env
+        long deleted = await env
             .RealtimeService.GetRepository<SFProjectUserConfig>()
             .DeleteAllAsync(p => p.Id == projectUserConfigId);
         Assert.AreEqual(1, deleted);
